imp(cli): Refactor Git operations into dedicated module (#97)
* refactor git from github module

* add changelog entry
